pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ontology-audit-hub"
version = "0.1.0"
description = "Ontology-driven QA audit hub built with LangGraph and Pydantic."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "langgraph>=0.2.27,<0.3",
  "langgraph-checkpoint-sqlite>=1.0.4,<2",
  "qdrant-client>=1.17",
  "fastapi>=0.115,<1",
  "python-multipart>=0.0.9,<1",
  "openai>=1.30,<3",
  "pydantic>=2.9,<3",
  "PyYAML>=6.0",
  "pypdf>=5,<6",
  "python-docx>=1.1,<2",
  "tiktoken>=0.9,<1",
  "typer>=0.24",
]

[project.optional-dependencies]
ai = ["pydantic-ai-slim[openai]>=1.76,<2"]
graph = ["neo4j>=6,<7"]
api = ["uvicorn>=0.30,<1"]
dev = ["httpx>=0.28", "mypy>=1.11", "pytest-cov>=5", "ruff>=0.8"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.hatch.build.targets.wheel]
packages = ["src/ontology_audit_hub"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008", "E501"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_optional = true
ignore_missing_imports = true
files = ["src"]
